Debugging kernel crash in TS combine

+++ Signal/Pulsar/dsp/CovarianceMatrixKernels.h
@@ -16,7 +16,7 @@ __global__ void outerProductKernel(float* result, unsigned int resultLength, flo
 __global__ void meanStokesKernel(float* d_amps, unsigned int ampsLength, const unsigned int* d_hits, unsigned int stokesLength);
 __global__ void applyScaleKernel(float* amps, unsigned int ampsLength, double scaleFactor);
 __global__ void genericAddKernel(unsigned int n, float* original, const float* add);
-__global__ void genericAddKernel(uint64_t n, float* original, const float* add);
+__global__ void genericAddKernel(unsigned long long int n, float* original, const float* add);
 __global__ void genericAddKernel(unsigned int n, unsigned int* original, const unsigned int* add);
 __global__ void genericSubtractionKernel(unsigned int n, float* original, const float* sub);
 __global__ void genericDivideKernel(unsigned int n, float* d_numerators, unsigned int denominator);
